[FIX] website_quote: quotation template with visible discount on pricelist
Steps to reproduce the bug:

Let's consider a product A with a price of 100
Let's consider a pricelist P with a visible discount of 50% on product A.
Let's consider a quotation template Q with one line with the product A with 50%
of discount.
Let's create a SO with pricelist P and set the quotation template Q on it.

Bug:

The unit price of A on SO is 50 instead of 100 and the visible discount is 50%
instead of 75%.

With the fix:

The unit price is 100 and the discount is 75%

opw:771996