[FIX] hr_expense: included tax and round globally
- Set 'Tax calculation rounding method' to 'Round Globally'
- Create a expense product with a tax of 20% included in the price
- Create a expense of 19.99
- Submit the expense in a expense sheet
- Validate the expense
- Post the expense
- Register a payment of 19.99

The expense is not set as paid.

The source of the issue is that the tax line is created without being
rounded. In the specific example, two lines are created with amounts:
- 16.658
- 3.33166667 (Tax)

This introduces rounding error in the computation of the matched
percentage, and the expense is not set as paid.

We force the rounding of the taxes since the 'Round Globally' option has
no impact on an expense report.

opw-773908