[FIX] mass_mailing: unsubscribe error if contacts
If the mass mailing was to mail contacts, this could cause an error if
the recipient unsubscribed. If the mass mailing was to mailing list,
unsubscribe would not work.

This change update unsubscribe to take into account changes in c4a2a6e
and fbed5d58c that allowed mailing list / mailing list contacts to be
used separately.

opw-770473
closes #20145