[FIX] web,report: display warning when report download is blocked
iOS devices need to open a window to allow downloading files.  This is
fine, unless the browser is setup to block the popups, which is quite
probably the default settings.  In that case, it just fails silently,
and the user has no idea of what just happened.

So, with this commit, we simply display a warning to let the user know
of what is actually going on.