[FIX] sale: product is editable after invoice, picking, timesheet
Since procurement removal introduced by commit ad66307dd19708f78b5fb3e1205c0f4f7a006648
product is editable once picking, timesheet,... are created.
However if the user modify the product it will not be reflected to
the related downstream object.

Previously it checks if a procurement order was linked to the sale line
and thus make it readonly.

This commit reproduce the old behavior by using a computed field that
can be override in sale_x modules in order to block the product
change for defined object.