[FIX] sale: pricelist price base
Fine tuning of afafb0addebc4457007296a7646b076b142c7e34

When setting a pricelist to show the public price discount, it ignored
the pricelist item's base field (if not 'Other Pricelist') and set it
by default on the lst_price instead.

This did not work when you based your discount on the standard_price
(= cost price) with public price discount.

Now it does.

opw:775275